Removed some redundant parameters, vectorised SUN shortened su2 multiply dag. Cleaned up some warnings flagged by Wextra.

+++ src/Headers/LU.h
@@ -31,18 +31,18 @@
 #else
 
 /**
-   @fn double complex LU_det( const int N , const GLU_complex U[ N*N ] )
+   @fn double complex LU_det( const size_t N , const GLU_complex U[ N*N ] )
    @brief computes the determinant of a matrix U
    @param N :: Side length of the square matrix
    @param U :: Matrix having its determinant taken
    @return the determinant
  */
 double complex
-LU_det( const int N , 
+LU_det( const size_t N , 
 	const GLU_complex U[ N*N ] ) ;
 
 /**
-   @fn double complex LU_det_overwrite( const int N , GLU_complex U[ N*N ] )
+   @fn double complex LU_det_overwrite( const size_t N , GLU_complex U[ N*N ] )
    @brief computes the determinant of a matrix U
    @param N :: Side length of the square matrix
    @param U :: Matrix having its determinant taken, overwritten
@@ -50,7 +50,7 @@ LU_det( const int N ,
    @warning overwrites space in U
  */
 double complex
-LU_det_overwrite( const int N , 
+LU_det_overwrite( const size_t N , 
 		  GLU_complex U[ N*N ] ) ;
