+++ src/Headers/LU_SSE.h
@@ -26,18 +26,18 @@
 #if ( defined HAVE_IMMINTRIN_H ) && !( defined SINGLE_PREC ) && (NC>3)
 
 /**
-   @fn double complex LU_det( const int N , const GLU_complex U[ N*N ] )
+   @fn double complex LU_det( const size_t N , const GLU_complex U[ N*N ] )
    @brief computes the determinant of a matrix U
    @param N :: Side length of the square matrix
    @param U :: Matrix having its determinant taken
    @return the determinant
  */
 double complex
-LU_det( const int N , 
+LU_det( const size_t N , 
 	const GLU_complex U[ N*N ] ) ;
 
 /**
-   @fn double complex LU_det_overwrite( const int N , GLU_complex U[ N*N ] )
+   @fn double complex LU_det_overwrite( const size_t N , GLU_complex U[ N*N ] )
    @brief computes the determinant of a matrix U
    @param N :: Side length of the square matrix
    @param U :: Matrix having its determinant taken, overwritten
@@ -45,7 +45,7 @@ LU_det( const int N ,
    @warning overwrites space in U
  */
 double complex
-LU_det_overwrite( const int N , 
+LU_det_overwrite( const size_t N , 
 		  GLU_complex U[ N*N ] ) ;
 
 #endif
