+++ src/Headers/plan_ffts.h
@@ -26,17 +26,17 @@
 #ifdef HAVE_FFTW3_H
 
 /**
-   @fn void clean_up_fftw( struct fftw_stuff FFTW , const int ARR_SIZE ) 
+   @fn void clean_up_fftw( struct fftw_stuff FFTW , const size_t ARR_SIZE ) 
    @brief memory deallocations for fftw_stuff struct allocated in create_plans_DFT()
    @param FFTW :: collection of temporary arrays
    @param ARR_SIZE :: corresponds to ARR_SIZE in create_plans_DFT
  */
 void
 clean_up_fftw( struct fftw_stuff FFTW ,
-	       const int ARR_SIZE ) ;
+	       const size_t ARR_SIZE ) ;
 
 /**
-   @fn void create_plans_DFT( struct fftw_stuff *FFTW , const size_t dims[ ND ] , const int ARR_SIZE , const int DIR )
+   @fn void create_plans_DFT( struct fftw_stuff *FFTW , const size_t dims[ ND ] , const int ARR_SIZE , const size_t DIR )
    @brief create the plans and allocate arrays for the DFT
    @param FFTW :: temporary arrays
    @param dims :: dimensions of the fft
@@ -46,8 +46,8 @@ clean_up_fftw( struct fftw_stuff FFTW ,
 void
 create_plans_DFT( struct fftw_stuff *FFTW ,
 		  const size_t dims[ ND ] ,
-		  const int ARR_SIZE ,
-		  const int DIR ) ;
+		  const size_t ARR_SIZE ,
+		  const size_t DIR ) ;
 
 /**
    @fn void small_clean_up_fftw( struct fftw_small_stuff FFTW ) 
@@ -58,7 +58,7 @@ void
 small_clean_up_fftw( struct fftw_small_stuff FFTW ) ;
 
 /**
-   @fn void small_create_plans_DFT( struct fftw_small_stuff *FFTW , const size_t dims[ ND ] , const int DIR )
+   @fn void small_create_plans_DFT( struct fftw_small_stuff *FFTW , const size_t dims[ ND ] , const size_t DIR )
    @brief create plans and allocate arrays of FFTW.*in etc
    @param FFTW :: temporary fft arrays
    @param dims :: dimensions of the FFT in GLU order
@@ -67,7 +67,7 @@ small_clean_up_fftw( struct fftw_small_stuff FFTW ) ;
 void
 small_create_plans_DFT( struct fftw_small_stuff *FFTW ,
 			const size_t dims[ ND ] ,
-			const int DIR ) ;
+			const size_t DIR ) ;
 
 #endif // HAVE_FFTW_H
 #endif
